refactor view parts directory
