+++ demo/src/app/view/parts/demo-parts-center/demo-parts-center.scss
