+++ demo/src/app/view/parts/demo-parts-dialog/demo-parts-dialog.scss
