+++ demo/src/app/view/parts/demo-parts-log/demo-parts-log.html
