+++ demo/src/app/view/parts/demo-parts-select/demo-parts-select.scss
